move to main grafan UI component
